pyproject.toml
[project]
name = "exif-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "pathlib>=1.0.1",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "streamlit>=1.45.1",
]
